% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqgofit.h.R
\name{chisqgofit}
\alias{chisqgofit}
\title{Goodness-of-Fit or Independence (Chi-square Test)}
\usage{
chisqgofit(
  calculate = "selectpower",
  typeinput = "stdeff",
  w = 0.13,
  df = 1,
  nrow = 2,
  ncol = 2,
  power = 0.8,
  alpha = 0.05,
  n = 200
)
}
\arguments{
\item{calculate}{.}

\item{typeinput}{.}

\item{w}{.}

\item{df}{.}

\item{nrow}{.}

\item{ncol}{.}

\item{power}{.}

\item{alpha}{.}

\item{n}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Goodness-of-Fit or Independence (Chi-square Test)
}
