% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqtest.h.R
\name{chisqtest}
\alias{chisqtest}
\title{Statistical Power for the Generic Chi-square Test}
\usage{
chisqtest(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Statistical Power for the Generic Chi-square Test
}
