% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicoPath-package.R
\docType{package}
\name{ClinicoPath-package}
\alias{ClinicoPath-package}
\alias{_PACKAGE}
\alias{ClinicoPath}
\title{ClinicoPath: Analysis for Clinicopathological Research}
\description{
`ClinicoPath` ClinicoPath help researchers to generate natural
language summaries of their dataset, generate cross tables with statistical
tests, and survival analysis with survival tables, survival plots,
and natural language summaries.


For more documentation, see the
\href{https://sbalci.github.io/ClinicoPathJamoviModule/}{Website}.
}
\details{
\code{ClinicoPath}
}
