% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancova.h.R
\name{fancova}
\alias{fancova}
\title{Analysis of (Co)Variance (F test)}
\usage{
fancova(
  calculate = "selectpower",
  eta2 = 0.01,
  f2 = 0.01,
  covadj = FALSE,
  ncov = 1,
  nway = "1",
  nlevelsA = 2,
  nlevelsB = 2,
  nlevelsC = 2,
  power = 0.8,
  alpha = 0.05,
  n = 200
)
}
\arguments{
\item{calculate}{.}

\item{eta2}{.}

\item{f2}{.}

\item{covadj}{.}

\item{ncov}{.}

\item{nway}{.}

\item{nlevelsA}{.}

\item{nlevelsB}{.}

\item{nlevelsC}{.}

\item{power}{.}

\item{alpha}{.}

\item{n}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Analysis of (Co)Variance (F test)
}
