% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freg.h.R
\name{freg}
\alias{freg}
\title{Linear Regression R squared or R squared Difference (F Test)}
\usage{
freg(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Linear Regression R squared or R squared Difference (F Test)
}
