% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmanova.h.R
\name{frmanova}
\alias{frmanova}
\title{Repeated Measures Analysis of Variance (F test)}
\usage{
frmanova(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Repeated Measures Analysis of Variance (F test)
}
