% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftest.h.R
\name{ftest}
\alias{ftest}
\title{Statistical Power for the Generic F Test}
\usage{
ftest(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Statistical Power for the Generic F Test
}
