% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np2groups.h.R
\name{np2groups}
\alias{np2groups}
\title{Difference between Two Groups (Non-parametric Tests for Independent and Paired Samples)}
\usage{
np2groups(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Difference between Two Groups (Non-parametric Tests for Independent and Paired Samples)
}
