% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2means.h.R
\name{t2means}
\alias{t2means}
\title{Difference between Two Means (t or z Test for Independent or Paired Samples)}
\usage{
t2means(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Difference between Two Means (t or z Test for Independent or Paired Samples)
}
