% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2meansdep.h.R
\name{t2meansdep}
\alias{t2meansdep}
\title{Two Dependent Means (t Test)}
\usage{
t2meansdep(
  calculate = "selectpower",
  paired.r = 0.5,
  stdinput = FALSE,
  d = 0.2,
  mu1 = 0.2,
  mu2 = 0,
  sd1 = 1,
  sd2 = 1,
  margin = 0,
  power = 0.8,
  alpha = 0.05,
  alternative = "not equal",
  kappa = 1,
  n2 = 200
)
}
\arguments{
\item{calculate}{.}

\item{paired.r}{.}

\item{stdinput}{.}

\item{d}{.}

\item{mu1}{.}

\item{mu2}{.}

\item{sd1}{.}

\item{sd2}{.}

\item{margin}{.}

\item{power}{.}

\item{alpha}{.}

\item{alternative}{.}

\item{kappa}{.}

\item{n2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Two Dependent Means (t Test)
}
