% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmean.h.R
\name{tmean}
\alias{tmean}
\title{One Mean t Test}
\usage{
tmean(
  calculate = "selectpower",
  mu = 0.25,
  sd = 1,
  mu0 = 0,
  margin = 0,
  power = 0.8,
  alpha = 0.05,
  alternative = "not equal",
  n = 200
)
}
\arguments{
\item{calculate}{.}

\item{mu}{.}

\item{sd}{.}

\item{mu0}{.}

\item{margin}{.}

\item{power}{.}

\item{alpha}{.}

\item{alternative}{.}

\item{n}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
One Mean t Test
}
