% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treg.h.R
\name{treg}
\alias{treg}
\title{Linear Regression Coef (t Test)}
\usage{
treg(
  calculate = "selectpower",
  predictor = "continuous",
  p = 0.5,
  stdinput = FALSE,
  beta1 = 0.25,
  beta0 = 0,
  margin = 0,
  sdx = 1,
  sdy = 1,
  k = 1,
  r2 = 0,
  power = 0.8,
  alpha = 0.05,
  alternative = "not equal",
  n = 200
)
}
\arguments{
\item{calculate}{.}

\item{predictor}{.}

\item{p}{.}

\item{stdinput}{.}

\item{beta1}{.}

\item{beta0}{.}

\item{margin}{.}

\item{sdx}{.}

\item{sdy}{.}

\item{k}{.}

\item{r2}{.}

\item{power}{.}

\item{alpha}{.}

\item{alternative}{.}

\item{n}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Linear Regression Coef (t Test)
}
