% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.h.R
\name{ttest}
\alias{ttest}
\title{Statistical Power for the Generic t Test}
\usage{
ttest(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Statistical Power for the Generic t Test
}
