% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2corrs.h.R
\name{z2corrs}
\alias{z2corrs}
\title{Difference between Two Correlations (Independent Samples z Test)}
\usage{
z2corrs(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Difference between Two Correlations (Independent Samples z Test)
}
