% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcorr.h.R
\name{zcorr}
\alias{zcorr}
\title{One Correlation against a Constant (One Sample z Test)}
\usage{
zcorr(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
One Correlation against a Constant (One Sample z Test)
}
