% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmed.h.R
\name{zmed}
\alias{zmed}
\title{Indirect Effect in Mediation Analysis (z, Joint, and Monte Carlo Tests)}
\usage{
zmed(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Indirect Effect in Mediation Analysis (z, Joint, and Monte Carlo Tests)
}
