% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpoisreg.h.R
\name{zpoisreg}
\alias{zpoisreg}
\title{Poisson Regression Single Coefficient (Large Sample Approx. Wald’s z Test)}
\usage{
zpoisreg(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Poisson Regression Single Coefficient (Large Sample Approx. Wald’s z Test)
}
