% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zprop.h.R
\name{zprop}
\alias{zprop}
\title{One Proportion z test}
\usage{
zprop(calculate = "selectpower", p = 0.4, p0 = 0.5)
}
\arguments{
\item{calculate}{.}

\item{p}{.}

\item{p0}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
One Proportion z test
}
