% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest.h.R
\name{ztest}
\alias{ztest}
\title{Statistical Power for the Generic z Test}
\usage{
ztest(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Statistical Power for the Generic z Test
}
