% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqgofit.h.R
\name{chisqgofit}
\alias{chisqgofit}
\title{Goodness-of-Fit or Independence (Chi-square Test)}
\usage{
chisqgofit(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Goodness-of-Fit or Independence (Chi-square Test)
}
