% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqgofit.h.R
\name{chisqgofit}
\alias{chisqgofit}
\title{Goodness-of-Fit or Independence (Chi-square Test)}
\usage{
chisqgofit(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Goodness-of-Fit or Independence (Chi-square Test)
}
