% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancova.h.R
\name{fancova}
\alias{fancova}
\title{Analysis of (Co)Variance (F test)}
\usage{
fancova(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Analysis of (Co)Variance (F test)
}
