% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancova.h.R
\name{fancova}
\alias{fancova}
\title{Analysis of (Co)Variance (F test)}
\usage{
fancova(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Analysis of (Co)Variance (F test)
}
