% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freg.h.R
\name{freg}
\alias{freg}
\title{Linear Regression R squared or R squared Difference (F Test)}
\usage{
freg(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Linear Regression R squared or R squared Difference (F Test)
}
