% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freg.h.R
\name{freg}
\alias{freg}
\title{Linear Regression R squared or R squared Difference (F Test)}
\usage{
freg(
  calculate = "selectpower",
  regmod = "rsqdev",
  r2 = 0.01,
  f2 = 0.01,
  k = 1,
  m = 1,
  power = 0.8,
  alpha = 0.05,
  n = 200
)
}
\arguments{
\item{calculate}{.}

\item{regmod}{.}

\item{r2}{.}

\item{f2}{.}

\item{k}{.}

\item{m}{.}

\item{power}{.}

\item{alpha}{.}

\item{n}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Linear Regression R squared or R squared Difference (F Test)
}
