% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmanova.h.R
\name{frmanova}
\alias{frmanova}
\title{Repeated Measures Analysis of Variance (F test)}
\usage{
frmanova(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Repeated Measures Analysis of Variance (F test)
}
