% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftest.h.R
\name{ftest}
\alias{ftest}
\title{Statistical Power for the Generic F Test}
\usage{
ftest(calculate = "selectpower", ncp = 1.96, df1 = 2, df2 = 199, alpha = 0.05)
}
\arguments{
\item{calculate}{.}

\item{ncp}{.}

\item{df1}{.}

\item{df2}{.}

\item{alpha}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Statistical Power for the Generic F Test
}
