% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np2groups.h.R
\name{np2groups}
\alias{np2groups}
\title{Difference between Two Groups (Non-parametric Tests for Independent and Paired Samples)}
\usage{
np2groups(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Difference between Two Groups (Non-parametric Tests for Independent and Paired Samples)
}
