% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2means.h.R
\name{t2means}
\alias{t2means}
\title{Difference between Two Means (t or z Test for Independent or Paired Samples)}
\usage{
t2means(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Difference between Two Means (t or z Test for Independent or Paired Samples)
}
