% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmean.h.R
\name{tmean}
\alias{tmean}
\title{One Mean against a Constant (z or t Test)}
\usage{
tmean(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
One Mean against a Constant (z or t Test)
}
