% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treg.h.R
\name{treg}
\alias{treg}
\title{Linear Regression Single Coefficient (t or z Test)}
\usage{
treg(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Linear Regression Single Coefficient (t or z Test)
}
