% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.h.R
\name{ttest}
\alias{ttest}
\title{Statistical Power for the Generic t Test}
\usage{
ttest(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Statistical Power for the Generic t Test
}
