% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2corrs.h.R
\name{z2corrs}
\alias{z2corrs}
\title{Difference between Two Correlations (Independent Samples z Test)}
\usage{
z2corrs(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Difference between Two Correlations (Independent Samples z Test)
}
