% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2props.h.R
\name{z2props}
\alias{z2props}
\title{Difference between Two Proportions (z Test)}
\usage{
z2props(
  calculate = "selectpower",
  p1 = 0.4,
  p2 = 0.5,
  margin = 0,
  power = 0.8,
  alpha = 0.05,
  alternative = "not equal",
  arcsintrans = FALSE,
  kappa = 1,
  n2 = 200
)
}
\arguments{
\item{calculate}{.}

\item{p1}{.}

\item{p2}{.}

\item{margin}{.}

\item{power}{.}

\item{alpha}{.}

\item{alternative}{.}

\item{arcsintrans}{.}

\item{kappa}{.}

\item{n2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Difference between Two Proportions (z Test)
}
