% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2props.h.R
\name{z2props}
\alias{z2props}
\title{Difference between Two Proportions (z Test)}
\usage{
z2props(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Difference between Two Proportions (z Test)
}
