% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcorr.h.R
\name{zcorr}
\alias{zcorr}
\title{One Correlation against a Constant (One Sample z Test)}
\usage{
zcorr(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
One Correlation against a Constant (One Sample z Test)
}
