% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlogreg.h.R
\name{zlogreg}
\alias{zlogreg}
\title{Logistic Regression Coef (z Test)}
\usage{
zlogreg(
  calculate = "selectpower",
  typeinput = "prob",
  p0 = 0.5,
  p1 = 0.4,
  oddsratio = 0.67,
  beta1 = 0.25,
  beta0 = 0,
  r2otherx = 0,
  power = 0.8,
  alpha = 0.05,
  alternative = "not equal",
  method = "demidenko(vc)",
  distribution = "not equal",
  mean = 0,
  sd = 1,
  min = 0,
  max = 1,
  size = 1,
  prob = 0.5,
  meanlog = 0,
  sdlog = 1,
  lambda = 1,
  rate = 1,
  n = 200
)
}
\arguments{
\item{calculate}{.}

\item{typeinput}{.}

\item{p0}{.}

\item{p1}{.}

\item{oddsratio}{.}

\item{beta1}{.}

\item{beta0}{.}

\item{r2otherx}{.}

\item{power}{.}

\item{alpha}{.}

\item{alternative}{.}

\item{method}{.}

\item{distribution}{.}

\item{mean}{.}

\item{sd}{.}

\item{min}{.}

\item{max}{.}

\item{size}{.}

\item{prob}{.}

\item{meanlog}{.}

\item{sdlog}{.}

\item{lambda}{.}

\item{rate}{.}

\item{n}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Logistic Regression Coef (z Test)
}
