% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmed.h.R
\name{zmed}
\alias{zmed}
\title{Indirect Effect in Mediation Analysis (z, Joint, and Monte Carlo Tests)}
\usage{
zmed(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Indirect Effect in Mediation Analysis (z, Joint, and Monte Carlo Tests)
}
