% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmed.h.R
\name{zmed}
\alias{zmed}
\title{Mediation (Sobel's z Test)}
\usage{
zmed(
  calculate = "selectpower",
  predictor = "continuous",
  stdinput = FALSE,
  p = 0.5,
  a = 0.25,
  b = 0,
  cp = 0,
  sdy = 1,
  sdx = 1,
  sdm = 1,
  r2y = 0,
  power = 0.8,
  alpha = 0.05,
  alternative = "not equal",
  n = 200
)
}
\arguments{
\item{calculate}{.}

\item{predictor}{.}

\item{stdinput}{.}

\item{p}{.}

\item{a}{.}

\item{b}{.}

\item{cp}{.}

\item{sdy}{.}

\item{sdx}{.}

\item{sdm}{.}

\item{r2y}{.}

\item{power}{.}

\item{alpha}{.}

\item{alternative}{.}

\item{n}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Mediation (Sobel's z Test)
}
