% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpoisreg.h.R
\name{zpoisreg}
\alias{zpoisreg}
\title{Poisson Regression Single Coefficient (Large Sample Approx. Wald’s z Test)}
\usage{
zpoisreg(calculate = "selectpower")
}
\arguments{
\item{calculate}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Poisson Regression Single Coefficient (Large Sample Approx. Wald’s z Test)
}
