% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zprop.h.R
\name{zprop}
\alias{zprop}
\title{One Proportion z Test}
\usage{
zprop(
  calculate = "selectpower",
  p = 0.4,
  p0 = 0.5,
  margin = 0,
  power = 0.8,
  alpha = 0.05,
  alternative = "not equal",
  arcsin.trans = FALSE,
  n = 200
)
}
\arguments{
\item{calculate}{.}

\item{p}{.}

\item{p0}{.}

\item{margin}{.}

\item{power}{.}

\item{alpha}{.}

\item{alternative}{.}

\item{arcsin.trans}{.}

\item{n}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
One Proportion z Test
}
