% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest.h.R
\name{ztest}
\alias{ztest}
\title{Statistical Power for the Generic z Test}
\usage{
ztest(
  calculate = "selectpower",
  ncp = 1.96,
  alpha = 0.05,
  alternative = "not equal"
)
}
\arguments{
\item{calculate}{.}

\item{ncp}{.}

\item{alpha}{.}

\item{alternative}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Statistical Power for the Generic z Test
}
